# docs/apispecs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kasir App",
        "description": "Kasir App API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/v1"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Product API"
                ],
                "description": "Get Products",
                "summary": "Get Products",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Products Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "product_type": {
                                                        "type": "string"
                                                    },
                                                    "stock": {
                                                        "type": "number"
                                                    },
                                                    "code_product": {
                                                        "type": "string"
                                                    },
                                                    "base_price": {
                                                        "type": "number"
                                                    },
                                                    "selling_price": {
                                                        "type": "number"
                                                    },
                                                    "category_id": {
                                                        "type": "number"
                                                    },
                                                    "minimum_stock": {
                                                        "type": "number"
                                                    },
                                                    "shelf": {
                                                        "type": "string"
                                                    },
                                                    "weight": {
                                                        "type": "number"
                                                    },
                                                    "discount": {
                                                        "type": "number"
                                                    },
                                                    "information": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product API"
                ],
                "description": "Create Product",
                "summary": "Create Product",
                "operationId": "createProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "product_type": {
                                        "type": "string"
                                    },
                                    "stock": {
                                        "type": "number"
                                    },
                                    "code_product": {
                                        "type": "string"
                                    },
                                    "base_price": {
                                        "type": "number"
                                    },
                                    "selling_price": {
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "type": "number"
                                    },
                                    "minimum_stock": {
                                        "type": "number"
                                    },
                                    "shelf": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "number"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "information": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Product Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/products_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/products/{product_id}": {
            "get": {
                "tags": [
                    "Product API"
                ],
                "description": "Get Product",
                "summary": "Get Product",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/products_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product API"
                ],
                "description": "Update Product",
                "summary": "Update Product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "product_type": {
                                        "type": "string"
                                    },
                                    "stock": {
                                        "type": "number"
                                    },
                                    "code_product": {
                                        "type": "string"
                                    },
                                    "base_price": {
                                        "type": "number"
                                    },
                                    "weight": {
                                        "type": "number"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "information": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Product Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "product_type": {
                                                        "type": "string"
                                                    },
                                                    "stock": {
                                                        "type": "number"
                                                    },
                                                    "code_product": {
                                                        "type": "string"
                                                    },
                                                    "base_price": {
                                                        "type": "number"
                                                    },
                                                    "selling_price": {
                                                        "type": "number"
                                                    },
                                                    "category_id": {
                                                        "type": "number"
                                                    },
                                                    "minimum_stock": {
                                                        "type": "number"
                                                    },
                                                    "shelf": {
                                                        "type": "string"
                                                    },
                                                    "weight": {
                                                        "type": "number"
                                                    },
                                                    "discount": {
                                                        "type": "number"
                                                    },
                                                    "information": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }    
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product API"
                ],
                "description": "Delete Product",
                "summary": "Delete Product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "message": {
                                                        "description": "Product Deleted",
                                                        "type": "string"
                                                    }
                                                }
                                             }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
            "schemas": {
...
                "products_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "name": {
                            "type": "string"
                        },
                        "product_type": {
                            "type": "string"
                        },
                        "stock": {
                            "type": "number"
                        },
                        "code_product": {
                            "type": "number"
                        },
                        "base_price": {
                            "type": "number"
                        },
                        "selling_price": {
                            "type": "number"
                        },
                        "category_id": {
                            "type": "number"
                        },
                        "minimum_stock": {
                            "type": "number"
                        },
                        "shelf": {
                            "type": "string"
                        },
                        "image": {
                            "type": "string"
                        },
                        "weight": {
                            "type": "number"
                        },
                        "discount": {
                            "type": "number"
                        },
                        "information": {
                            "type": "string"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
...
            },
            "securitySchemes": {
                    "BearerAuth": {
                        "type": "http",
                        "scheme": "bearer",
                        "bearerFormat": "JWT"
                    }
                }
        }
}